+++ CHANGELOG.md
@@ -275,7 +275,9 @@
 ## dev
 
 - Improve Deepparse server error handling and error output
-- Remove deprecated argument ``saving_dir`` in ``download_fasttext_magnitude_embeddings``
-  and ``download_fasttext_embeddings`` functions
-- Add offline argument to remove verification of latest version.
+- Remove deprecated argument `saving_dir` in `download_fasttext_magnitude_embeddings`
+  and `download_fasttext_embeddings` functions
+- Add offline argument to remove verification of the latest version
+- Bug-fix cache handling in download model
+- Add `download_models`` CLI function